+++ mybatis-r2dbc/src/main/java/pro/chenggang/project/reactive/mybatis/support/r2dbc/executor/ReactiveMybatisExecutor.java
@@ -35,7 +35,7 @@ public interface ReactiveMybatisExecutor {
      * @param parameter       the parameter
      * @return mono
      */
-    Mono<Integer> update(MappedStatement mappedStatement, Object parameter);
+    Mono<Long> update(MappedStatement mappedStatement, Object parameter);
 
     /**
      * execute query
